chore: move surviving pkg/ packages into internal/queueing
