+++ docs/developer-guide/development.md
@@ -319,7 +319,7 @@ kubectl logs -n workload-variant-autoscaler-system \
 
 ### Modifying Optimization Logic
 
-1. Update code in `pkg/solver/` or `pkg/analyzer/`
+1. Update code in `internal/engines/pipeline/` or `internal/queueing/analyzer/`
 2. Add/update unit tests
 3. Run `make test`
 4. Update design documentation if algorithm changes
