+++ internal/queueing/analyzer/mm1kmodel.go
