+++ internal/queueing/analyzer/utils.go
