+++ internal/queueing/analyzer/utils_test.go
